requirements.txt
flask-bcrypt == 1.0.1
Flask == 2.1.2
Flask_Cors == 3.0.10
Flask_JWT_Extended == 4.4.4
flask_mongoengine == 1.0.0
flask_restx == 0.5.1
mongoengine == 0.24.2
pydantic == 2.27.1
pymongo == 4.2.0
python-dotenv == 0.21.0
Werkzeug == 2.1.2
python-json-logger == 2.0.4

//previous
//pydantic  1.10.2